chore: cypress tests for adding map layers (#1738)
Test adding map layers. It does not yet test the actual rendered map.

The files in the /elements folder are a first attempt at page objects. They are classes that hide complexity of interacting with the page so that the actual tests in the /integration folder read nicely. Most methods in the page objects return this so that the tests can chain the methods.