fix: parse date string to array to get correct date in all timezones (#1937)
fixes: https://jira.dhis2.org/browse/DHIS2-12169

If a string with this format "2018-10-01" is passed to javascript Date, the assumed time is "00:00" and the returned date will be according to the user's timezone. So if you are in <= GMT-1, then the resulting date will actually be the previous day. To avoid this, the date string is converted to an array, and Date converts this correctly.

The screenshot shows the Layer card with the incorrect date shown (note that the analytics request uses the correct date)