fix: parse date string to array to get correct date in all timezones (#1988)
fixes: https://jira.dhis2.org/browse/DHIS2-12169

If a string with this format "2018-10-01" is passed to javascript Date, the assumed time is "00:00" and the returned date will be according to the user's timezone. So if you are in <= GMT-1, then the resulting date shown in the card will actually be the previous day (though the analytics request does use the correct date, so the data is correct).

The fix is to convert the date string to an array, and Date converts this correctly.