feat: use default basemap that is configured in system settings [DHIS2-12155] (#1984)
keyDefaultBaseMap is fetched as part of the system settings. If no value found (there is no default value set on the server side), then the app and plugin use osmLight as the default.

Refactor: All map config modifications ( getBasemap,upgradeGisAppLayers and renameBoundaryToOrgUnit) are moved to one module (getMigratedMapConfig.js) and run in the same order as before as soon as the map is fetched.

FileMenu: former epics (loadFavorite, saveFavorite, saveNewFavorite) have been deleted, and those functions are now in FileMenu and use the dataEngine (openMap, saveMap, saveAsNewMap). The new functions in FileMenu do the exact same work as what was previously in the epics. Errors are caught and the user is shown an alert bar.