chore: add shim to provide baseUrl to apiFetch (#2446)
apiFetch currently uses d2 to get the baseUrl. This PR is part of removing d2 from maps, which is also good for the migration to platform.

apiFetchWithBaseUrl is a copy of apiFetch except that it uses the passed in baseUrl instead of baseUrl from d2. Eventually both these functions will go away, but for now i wanted to keep them separate since apiFetch is dependent on d2