feat: upgrade toolbar (DHIS2-15667) (#2936)
Changes in the following files are functional changes needed for the new Analytics toolbar:

src/components/app/AppMenu.js - add the HoverMenuBar component and onFileMenuAction callback, which updates the interpretations panel with changes (e.g. sharing, renaming, translating) if it is open
AppLayout - onFileMenuAction and the interpretationRenderId.
FileMenu - added onFileMenuAction
Interpretations, InterpretationsPanel: - added renderId which triggers refetching the map info if it was changed by a file menu action
InterpretationsToggle: use the new InterpretationsAndDetailsToggler from Analytics
DownloadButton: needed different style to look nice with the new toolbar
AddLayerButton: styled to fit into the new toolbar
The new toolbar is a different height than the old one, which caused a bunch of layout issues due to the absolute positioning. Switched to flex where it made sense, and css changes accordingly.

Only change to DownloadSettings is the addition of a wrapping div <div className={styles.downloadSettingsPanel}>
MapPosition: resize due to open/close of the Layers or Details panel has a timeout added, since there is now a 100ms transition when open/close those Drawers