fix: loader alert messages were being duplicated (#2796)
Fixes: https://dhis2.atlassian.net/browse/DHIS2-15438

The alerts were duplicated because there was both an internal alert stack/bar
system as well as the built in one from app-runtime. Solution was to remove
the internal alert stack/bar. This meant that the alerts coming from the layer
loaders had to be refactored so that loader alerts are no longer are added to redux.
Instead the component LayerLoaders use the app-runtime hook (useAlert).

Other changes:

skipPluginLogic: true has been added to d2.config.js because of the cli-app-scripts upgrade.
Previously we had set a resolution for the dependency to workaround this bug: https://dhis2.atlassian.net/browse/DHIS2-16916.
Now, we can remove the resolution as long as we include this config.