fix: plugin crashes for Earth Engine layers in the dashboard (#3302)
Maps plugin was crashing in the dashboard for EE maps due to a missing filters property.
The filters property for earth engine layers was recently removed because of a naming
conflict between the saved map object (for which the filters prop will always be an empty array)
and the earth engine layer config, which can contain a filters property for other purposes.

The problem arose in the plugin because in the dashboard a user can apply filters to the
visualizations, ncluding maps.
So if the filters property has been removed (as is the case for EE layers), then
the crash would occur. Since dashboard filtering has no effect on ee layers anyway, there is
no need to even check the filters property when determining whether to reload the map with
a dashboard filter applied.