chore: upgrade cli-app-scripts to fix AppHub publishing failure (#3393)
After the fix for Laos language support to glyphs was added (DHIS2-16395), 
the step to publish the build to AppHub started failing. This is because the
build bundle size exceeded that allowed by axios, which is used for posting
the bundle to AppHub. The bundle size max was fixed in cli-app-scripts 11.7.5.

This commit also sets @dhis2/ui as a direct dependency so that this app controls
which version is installed instead of depending on cli-app-scripts->app-shell
determining the version of @dhis2/ui.