fix: allow editing invalid layers (#3396)
Fixes DHIS2-18157

If a user configures an invalid layer, it should be possible for them to edit it.
For geojson url layers this isn't useful since there's nothing the user can do to fix an
invalid geojson url layer. But for thematic layers, the user should be able to
fix whatever is wrong most of the time.