fix: prevent DE or Attr used for styling from disappearing from popup [DHIS2-17711] (#3481)
name of styleDataItem is not saved with map, get the name with loadDisplayElements
just as other data elements marked to be displayed.
Handle scenario where an attribute is used for styling by fetching program TEA metadata.
Rename "Style by data element" to "Style by data item" to clarify the possibilities.