fix: do not overwrite newer changes by another user when renaming [DHIS2-19479] [DHIS2-19506] (#3520)
Implements:

DHIS2-19479
DHIS2-19506

Description
When fetching the map, fetch both name and displayName because both are needed for
different purposes. When renaming, the original name should be used because that is the
value that will be changed. But the map name, shown above the map canvas,
should use displayName so that the correct name translation is used when a different
database language is set.

When renaming a map, the original map needs to be refetched so as to not overwrite
any newer changes on the database. Because of this fetch just before PUT, it is no longer
necessary to keep a copy of the original map in redux, so that has been removed.
One side-effect of this fetch is that maps saved in an older format will get migrated
to the latest format.