chore: remove d2 [DHIS2-18205] (#3522)
Implements DHIS2-18205

Description
Replace use of d2 with the platform app-runtime and our own AnalyticsEngine.
AnalyticsEngine was built to be nearly identical to d2's api.

Some notes:

in MapContainer and thematicLoader, userOrgUnit property hasn't been in use since
dashboard filtering was refactored back in 2020. Therefore, code related to userOrgUnit in
maps is obsolete and is removed.

The var name nameProperty is used inconsistently in the app. Sometimes it means
"name/shortName", other times it means "displayName/displayShortName". To start to
make this more consistent, we now use the property name that is the actual system setting property "keyAnalysisDisplayProperty" to mean "name/shortName", instead of "nameProperty".
In the future we should decide on a consistent prop name for "displayName/displayShortName".